Force short

+++ src/qz/communication/H4J_HidUtilities.java
@@ -37,7 +37,7 @@ public class H4J_HidUtilities {
 
             deviceJSON.put("vendorId", UsbUtil.toHexString(device.getVendorId()))
                     .put("productId", UsbUtil.toHexString(device.getProductId()))
-                    .put("usagePage", UsbUtil.toHexString(device.getUsagePage()))
+                    .put("usagePage", UsbUtil.toHexString((short)device.getUsagePage()))
                     .put("serial", device.getSerialNumber())
                     .put("manufacturer", device.getManufacturer())
                     .put("product", device.getProduct());
@@ -60,7 +60,7 @@ public class H4J_HidUtilities {
         List<HidDevice> devices = getHidDevices();
         for(HidDevice device : devices) {
             if (device.isVidPidSerial(vendorId, productId, serial)
-                    && (usagePage == null || usagePage == device.getUsagePage())) {
+                    && (usagePage == null || usagePage == (short)device.getUsagePage())) {
                 return device;
             }
         }
